+++ src/TableEditColumn.js
@@ -9,14 +9,19 @@ class TableEditColumn extends Component {
     super(props);
     this.timeouteClear = 0;
     const { fieldValue, row, className } = this.props;
+    this.focusInEditor = this.focusInEditor.bind(this);
     this.state = {
       shakeEditor: false,
       className: typeof className === 'function' ? className(fieldValue, row) : className
     };
   }
 
+  valueShortCircuit(value) {
+    return value === null || typeof value === 'undefined' ? '' : value;
+  }
+
   handleKeyPress = e => {
-    if (e.keyCode === 13) {
+    if (e.keyCode === 13 || e.keyCode === 9) {
       // Pressed ENTER
       const value = e.currentTarget.type === 'checkbox' ?
                       this._getCheckBoxValue(e) : e.currentTarget.value;
@@ -24,13 +29,15 @@ class TableEditColumn extends Component {
       if (!this.validator(value)) {
         return;
       }
-      this.props.completeEdit(value, this.props.rowIndex, this.props.colIndex);
+      if (e.keyCode === 13) {
+        this.props.completeEdit(value, this.props.rowIndex, this.props.colIndex);
+      } else {
+        this.props.onTab(this.props.rowIndex + 1, this.props.colIndex + 1, 'tab', e);
+        e.preventDefault();
+      }
     } else if (e.keyCode === 27) {
       this.props.completeEdit(
         null, this.props.rowIndex, this.props.colIndex);
-    } else if (e.keyCode === 9) {
-      this.props.onTab(this.props.rowIndex + 1, this.props.colIndex + 1, 'tab', e);
-      e.preventDefault();
     } else if (e.type === 'click' && !this.props.blurToSave) {  // textarea click save button
       const value = e.target.parentElement.firstChild.value;
       if (!this.validator(value)) {
@@ -47,7 +54,7 @@ class TableEditColumn extends Component {
       const value = e.currentTarget.type === 'checkbox' ?
                       this._getCheckBoxValue(e) : e.currentTarget.value;
       if (!this.validator(value)) {
-        return;
+        return false;
       }
       this.props.completeEdit(
           value, this.props.rowIndex, this.props.colIndex);
@@ -67,7 +74,6 @@ class TableEditColumn extends Component {
     const ts = this;
     let valid = true;
     if (ts.props.editable.validator) {
-      const input = ts.refs.inputRef;
       const checkVal = ts.props.editable.validator(value, this.props.row);
       const responseType = typeof checkVal;
       if (responseType !== 'object' && checkVal !== true) {
@@ -90,7 +96,7 @@ class TableEditColumn extends Component {
         ts.timeouteClear = setTimeout(() => {
           ts.setState({ shakeEditor: false });
         }, 300);
-        input.focus();
+        this.focusInEditor();
         return valid;
       }
     }
@@ -117,7 +123,7 @@ class TableEditColumn extends Component {
   }
 
   componentDidMount() {
-    this.refs.inputRef.focus();
+    this.focusInEditor();
     const dom = ReactDOM.findDOMNode(this);
     if (this.props.isFocus) {
       dom.focus();
@@ -137,6 +143,12 @@ class TableEditColumn extends Component {
 
   componentWillUnmount() {
     this.clearTimeout();
+  }
+
+  focusInEditor() {
+    if (typeof this.refs.inputRef.focus === 'function') {
+      this.refs.inputRef.focus();
+    }
   }
 
   handleClick = e => {
@@ -167,18 +179,18 @@ class TableEditColumn extends Component {
     editable.placeholder && (attr.placeholder = editable.placeholder);
 
     const editorClass = classSet({ 'animated': shakeEditor, 'shake': shakeEditor });
+    fieldValue = fieldValue === 0 ? '0' : fieldValue;
     let cellEditor;
     if (customEditor) {
       const customEditorProps = {
         row,
         ...attr,
-        defaultValue: fieldValue || '',
+        defaultValue: this.valueShortCircuit(fieldValue),
         ...customEditor.customEditorParameters
       };
       cellEditor = customEditor.getElement(this.handleCustomUpdate, customEditorProps);
     } else {
-      fieldValue = fieldValue === 0 ? '0' : fieldValue;
-      cellEditor = editor(editable, attr, format, editorClass, fieldValue || '');
+      cellEditor = editor(editable, attr, format, editorClass, this.valueShortCircuit(fieldValue));
     }
 
     if (isFocus) {